+++ graal/com.oracle.graal.nodes/src/com/oracle/graal/nodes/calc/SignExtendNode.java
@@ -26,6 +26,7 @@ import com.oracle.graal.api.meta.*;
 import com.oracle.graal.graph.*;
 import com.oracle.graal.graph.spi.*;
 import com.oracle.graal.nodes.*;
+import com.oracle.graal.nodes.extended.*;
 import com.oracle.graal.nodes.spi.*;
 import com.oracle.graal.nodes.type.*;
 
@@ -106,5 +107,15 @@ public class SignExtendNode extends IntegerConvertNode {
     @Override
     public void generate(ArithmeticLIRGenerator gen) {
         gen.setResult(this, gen.emitSignExtend(gen.operand(getInput()), getInputBits(), getResultBits()));
+    }
+
+    @Override
+    public boolean generate(MemoryArithmeticLIRLowerer gen, Access access) {
+        assert !access.nullCheckLocation().getValueKind().isUnsigned() : "can't sign extend unsigned value";
+        Value result = gen.emitSignExtendMemory(access, access.nullCheckLocation().getValueKind().getBitCount(), getResultBits());
+        if (result != null) {
+            gen.setResult(this, result);
+        }
+        return result != null;
     }
 }